Pull export code out into its own file
It's shared between verify-source and import.
